+++ src/main/java/com/assignment/TicTacToe/Game/GameRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 import java.util.Optional;
 @Repository
-public interface GameRepository extends CrudRepository<Game, Long> {
-    List<Game> findByPlayer1IdOrPlayer2Id(long player1Id, long player2Id);
-    Optional<Game> findByPlayer1IdAndPlayer2Id(long player1Id, long player2Id);
+public interface GameRepository extends CrudRepository<Game, Integer> {
+    List<Game> findByPlayer1IdOrPlayer2Id(int player1Id, int player2Id);
+    Optional<Game> findByPlayer1IdAndPlayer2Id(int player1Id, int player2Id);
 }
